Move SearchQueryArgs to own file.
